requirements.txt
# ═══════════════════════════════════════════════════════════════
# Echo-Logic Hub — Python Dependencies
# ═══════════════════════════════════════════════════════════════
# Install: pip install -r requirements.txt
#
# For mock-mode testing (no GPU), only streamlit, python-dotenv,
# google-generativeai, and numpy are strictly required.
# ═══════════════════════════════════════════════════════════════

# --- Core UI ---
streamlit>=1.31.0

# --- LLM ---
google-generativeai>=0.5.0

# --- Configuration ---
python-dotenv>=1.0.0

# --- Audio ---
PyAudio>=0.2.14
numpy>=1.24.0
soundfile>=0.12.0

# --- NVIDIA NeMo (requires CUDA-capable GPU) ---
# Uncomment the line below when running with real STT models.
# Ensure your CUDA toolkit version matches your PyTorch build.
# nemo_toolkit[asr]>=2.0.0
